Avoid download samba4 release tarball if already available in current directory.

+++ script/installsamba4.sh
@@ -146,9 +146,11 @@ download() {
     fi
 
     echo "Step2: Fetching samba-$SAMBA4_RELEASE tarball"
-    rm -rf samba-$SAMBA4_RELEASE.tar.gz
-    wget http://us1.samba.org/samba/ftp/samba4/samba-$SAMBA4_RELEASE.tar.gz
-    error_check $? "Step1"
+    if ! test -e samba-$SAMBA4_RELEASE.tar.gz; then
+	rm -rf samba-$SAMBA4_RELEASE.tar.gz
+	wget http://us1.samba.org/samba/ftp/samba4/samba-$SAMBA4_RELEASE.tar.gz
+	error_check $? "Step1"
+    fi     
 
     echo "Step3: Extracting $SAMBA4_RELEASE"
     tar xzvf samba-$SAMBA4_RELEASE.tar.gz
